requirements.txt
mariadb==1.1.3
python-decouple==3.6
